修改逻辑

+++ src/com/ljscode/base/BaseReadUSBData.java
@@ -2,6 +2,6 @@ package com.ljscode.base;
 
 public interface BaseReadUSBData {
 
-    void ReadUSBData(double deg, double cylinder, double endFace);
+    void ReadUSBData(int deg, double cylinder, double endFace);
 
 }
